+++ src/r3_test_suite/src/lib.rs
@@ -270,6 +270,7 @@ pub mod kernel_tests {
         (mod semaphore_signal_and_dispatch {}, "semaphore_signal_and_dispatch"),
         (mod semaphore_timeout {}, "semaphore_timeout"),
         (mod startup_hook_disallowed_services {}, "startup_hook_disallowed_services"),
+        (mod startup_hook_misc {}, "startup_hook_misc"),
         (mod startup_hook_pend_interrupt {}, "startup_hook_pend_interrupt"),
         (mod startup_hook_priority {}, "startup_hook_priority"),
         (mod sync_mutex_lock_and_dispatch {}, "sync_mutex_lock_and_dispatch"),